third_party/WebKit/LayoutTests/css3/device-adapt/opera/descriptor-values-004-expected.txt


FAIL CSS Test: @viewport invalid zoom factor - drop negative factors assert_equals: expected 0 but got 1
